+++ src/idmatcherai/matcher.py
@@ -99,7 +99,7 @@ def load_identification_model(model_folder: Path):
     identification_network_params = NetworkParams(
         schedule=params["schedule"],
         n_classes=n_classes,
-        architecture="idCNN",
+        architecture="CNN",
         restore_folder=model_folder,
         model_name=params["model_name"],
         image_size=params["image_size"],