+++ src/idtrackerai/network/models.py
@@ -4,7 +4,7 @@ from typing import Sequence
 from torch import Tensor, nn
 
 
-class DCD(nn.Module):
+class CNN(nn.Module):
     def __init__(self, input_shape: Sequence[int], out_dim: int):
         super().__init__()
 
@@ -23,30 +23,7 @@ class DCD(nn.Module):
             nn.Linear(100, out_dim),
         )
 
-    def forward(self, x: Tensor):
-        return self.layers(x)
-
-
-class idCNN(nn.Module):
-    def __init__(self, input_shape: Sequence[int], out_dim: int):
-        super().__init__()
-
-        self.layers = nn.Sequential(
-            nn.Conv2d(input_shape[-1], 16, 5, padding=2),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(2, stride=2),
-            nn.Conv2d(16, 64, 5, padding=2),
-            nn.ReLU(inplace=True),
-            nn.MaxPool2d(2, stride=2),
-            nn.Conv2d(64, 100, 5, padding=2),
-            nn.ReLU(inplace=True),
-            nn.Flatten(),
-            nn.Linear(100 * (input_shape[1] // 4) ** 2, 100),
-            nn.ReLU(inplace=True),
-            nn.Linear(100, out_dim),
-        )
-
-    def forward(self, x: Tensor):
+    def forward(self, x: Tensor) -> Tensor:
         return self.layers(x)